requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
supabase==2.27.0
websockets>=15.0.0
email-validator>=2.0.0


